--- components/apimgt-extensions/org.wso2.carbon.apimgt.webapp.publisher/src/test/resources/carbon-home/repository/conf/apim-integration.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><!--
-/*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
- -->
-
-<APIMConfiguration>
-	<DCREndpoint>https://${iot.keymanager.host}:${iot.keymanager.https.port}/client-registration/v0.11/register</DCREndpoint>
-	<TokenEndpoint>https://${iot.gateway.host}:${iot.gateway.https.port}/token</TokenEndpoint>
-	<PublisherEndpoint>https://localhost:9445/api/am/publisher/v0.11</PublisherEndpoint>
-	<StoreEndpoint>https://${iot.apimstore.host}:${iot.apimstore.https.port}/api/am/store/v0.11</StoreEndpoint>
-
-	<DCREndpoint>https://localhost:9443/client-registration/v0.11/register</DCREndpoint>
-	<TokenEndpoint>https://localhost:9443/token</TokenEndpoint>
-	<PublisherEndpoint>https://localhost:9445/api/am/publisher/v0.11</PublisherEndpoint>
-	<StoreEndpoint>https://localhost:9445/api/am/store/v0.11</StoreEndpoint>
-
-	<Username>admin</Username>
-	<Password>admin</Password>
-</APIMConfiguration>
