--- components/apimgt-extensions/org.wso2.carbon.apimgt.webapp.publisher/src/test/resources/config/datasource/data-source-config.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-<!--
-  ~ Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-  ~
-  ~ WSO2 Inc. licenses this file to you under the Apache License,
-  ~ Version 2.0 (the "License"); you may not use this file except
-  ~ in compliance with the License.
-  ~ you may obtain a copy of the License at
-  ~
-  ~   http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing,
-  ~ software distributed under the License is distributed on an
-  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~ KIND, either express or implied.  See the License for the
-  ~ specific language governing permissions and limitations
-  ~ under the License.
-  -->
-
-<DataSourceConfig>
-    <Url>jdbc:h2:mem:cdm-test-db;DB_CLOSE_ON_EXIT=FALSE;MVCC=true</Url>
-    <DriverClassName>org.h2.Driver</DriverClassName>
-    <User>wso2carbon</User>
-    <Password>wso2carbon</Password>
-
-
-    <!-- For MySql -->
-
-    <!--<Url>jdbc:mysql://localhost:3306/WSO2CDM</Url>-->
-    <!--<DriverClassName>com.mysql.jdbc.Driver</DriverClassName>-->
-    <!--<User>root</User>-->
-    <!--<Password></Password>-->
-</DataSourceConfig>
