--- components/apimgt-extensions/org.wso2.carbon.apimgt.webapp.publisher/src/test/resources/config/schema/device-mgt-config-schema.xsd
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-  ~
-  ~ WSO2 Inc. licenses this file to you under the Apache License,
-  ~ Version 2.0 (the "License"); you may not use this file except
-  ~ in compliance with the License.
-  ~ you may obtain a copy of the License at
-  ~
-  ~   http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing,
-  ~ software distributed under the License is distributed on an
-  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~ KIND, either express or implied.  See the License for the
-  ~ specific language governing permissions and limitations
-  ~ under the License.
-  -->
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" elementFormDefault="qualified"
-           attributeFormDefault="unqualified">
-    <xs:element name="DeviceMgtConfiguration">
-        <xs:complexType>
-            <xs:sequence>
-                <xs:element name="ManagementRepository" minOccurs="1" maxOccurs="1" type="DataSourceConfigurationType"/>
-            </xs:sequence>
-        </xs:complexType>
-    </xs:element>
-
-    <xs:complexType name="JndiLookupDefinitionType">
-        <xs:sequence>
-            <xs:element name="Name" type="xs:string" minOccurs="1" maxOccurs="1"/>
-        </xs:sequence>
-    </xs:complexType>
-
-    <xs:complexType name="DataSourceConfigurationType">
-        <xs:sequence>
-            <xs:element name="JndiLookupDefinition" minOccurs="1" maxOccurs="1"
-                        type="JndiLookupDefinitionType"/>
-        </xs:sequence>
-    </xs:complexType>
-
-    <xs:complexType name="ManagementRepositoryType">
-        <xs:sequence>
-            <xs:element name="DataSourceConfiguration" minOccurs="1" maxOccurs="1"
-                        type="DataSourceConfigurationType"/>
-        </xs:sequence>
-    </xs:complexType>
-</xs:schema>
