--- components/apimgt-extensions/org.wso2.carbon.apimgt.webapp.publisher/src/test/resources/jwt.properties
@@ -1,57 +1,0 @@
-#
-# Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-#
-# WSO2 Inc. licenses this file to you under the Apache License,
-# Version 2.0 (the "License"); you may not use this file except
-# in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied. See the License for the
-# specific language governing permissions and limitations
-# under the License.
-#
-
-#issuer of the JWT
-iss=wso2.org/products/iot
-
-TokenEndpoint=https://${iot.gateway.host}:${iot.gateway.https.port}/token?tenantDomain=carbon.super
-
-#audience of JWT claim
-#comma seperated values
-aud=devicemgt
-
-#expiration time of JWT (number of minutes from the current time)
-exp=1000
-
-#issued at time of JWT (number of minutes from the current time)
-iat=0
-
-#nbf time of JWT (number of minutes from current time)
-nbf=0
-
-#skew between IDP and issuer(seconds)
-skew=0
-
-# JWT Id
-#jti=token123
-
-#KeyStore to cryptographic credentials
-#KeyStore=repository/resources/security/wso2carbon.jks
-
-#Password of the KeyStore
-#KeyStorePassword=wso2carbon
-
-#Alias of the SP's private key
-#PrivateKeyAlias=wso2carbon
-
-#Private key password to retrieve the private key used to sign
-#AuthnRequest and LogoutRequest messages
-#PrivateKeyPassword=wso2carbon
-
-#this will be used as the default IDP config if there isn't any config available for tenants.
-default-jwt-client=true
